- feat: add support for cohort renaming - feat: implement cohort selection
