+++ src/ui/SmallOverviewWrapper.tsx
@@ -88,6 +88,7 @@ export const SmallOverviewWrapper = ({ demo, setDemo, demoIndex, filteredSamples
         } else {
             setDoneGeneratingThumbnails(true);
         }
+        console.log('Filtered samples changed', filteredSamples);
     }, [demo, filteredSamples, thumbnailForceGenerate, generateThumbnails]);
 
     return filteredSamples.map((d, i) => (
